Added Main Menu and Credits Menu
Added a functional Main Menu and a semi functional Credits Menu. Main Menu has three buttons; Play, Credits, and Quit. When clicking on Play the scene transitions to the Tutorial Scene. Credits transitions the scene to the Credits Scene. Quit exits the build from executing. Credits Menu has a single Back button that transitions the scene to the MainMenu scene. There will be a scrolling list of Credits once their references are obtained.